# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciar dados da NASA APOD (Astronomy Picture of the Day)",
        "title": "AstroVista API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/apod": {
            "get": {
                "description": "Retorna a imagem astronômica do dia mais recente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APOD"
                ],
                "summary": "Obtém o APOD mais recente",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Apod"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Busca na API da NASA o APOD mais recente e adiciona ao banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APOD"
                ],
                "summary": "Adiciona novo APOD da NASA",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de API interno",
                        "name": "X-API-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/apod/{date}": {
            "get": {
                "description": "Retorna a imagem astronômica do dia para a data especificada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APOD"
                ],
                "summary": "Obtém um APOD por data específica",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"2023-01-15\"",
                        "description": "Data no formato YYYY-MM-DD",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Apod"
                        }
                    },
                    "400": {
                        "description": "Erro ao obter APOD",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.Apod": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "ID do MongoDB\nexample: 507f1f77bcf86cd799439011",
                    "type": "string"
                },
                "date": {
                    "description": "Data no formato string (ex: \"1995-06-16\")\nexample: 2023-01-15\nformat: date",
                    "type": "string"
                },
                "explanation": {
                    "description": "Explicação da imagem astronômica do dia\nexample: Uma bela nebulosa capturada pelo telescópio Hubble",
                    "type": "string"
                },
                "hdurl": {
                    "description": "URL da imagem em alta definição\nexample: https://apod.nasa.gov/apod/image/2301/M31_HubbleSpitzerGendler_960.jpg\nformat: uri",
                    "type": "string"
                },
                "media_type": {
                    "description": "Tipo de mídia (imagem ou vídeo)\nexample: image\nenum: image,video",
                    "type": "string"
                },
                "service_version": {
                    "description": "Versão do serviço da API\nexample: v1",
                    "type": "string"
                },
                "title": {
                    "description": "Título da imagem astronômica do dia\nexample: Galáxia de Andrômeda",
                    "type": "string"
                },
                "url": {
                    "description": "URL da imagem em resolução padrão\nexample: https://apod.nasa.gov/apod/image/2301/M31_HubbleSpitzerGendler_960.jpg\nformat: uri",
                    "type": "string"
                }
            }
        },
...
    }
}